requirements.txt
pytest==3.6.3
Flask==1.0.2
coverage==4.5.1
coveralls==1.3.0
python-dotenv==0.8.2
psycopg2-binary==2.7.5
Faker==0.8.17
py-bcrypt==0.4
PyJWT==1.6.4
